pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "document-intelligence-platform"
version = "0.3.0"
description = "Scalable end-to-end document intelligence platform with OCR, multi-type extraction, review workflow, webhooks, and observability."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  # Web
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "uvloop>=0.20.0; platform_system != 'Windows'",
  "httpx>=0.27.2",
  "python-multipart>=0.0.9",
  "openpyxl>=3.1.5",
  "anthropic>=0.45.2",
  # DB / ORM
  "sqlalchemy>=2.0.35",
  "psycopg[binary]>=3.2.1",
  "alembic>=1.13.0",
  # Config / validation
  "pydantic>=2.8.2",
  "pydantic-settings>=2.4.0",
  # Async task queue
  "celery[redis]>=5.4.0",
  "redis>=5.0.8",
  # Document processing
  "PyMuPDF>=1.24.10",
  "Pillow>=10.4.0",
  "numpy>=1.26.4",
  "pytesseract>=0.3.13",
  "pdfplumber>=0.11.4",
  # NLP / ML
  "spacy>=3.7.6",
  "python-dateutil>=2.9.0.post0",
  "rapidfuzz>=3.9.7",
  # S3 / cloud storage
  "boto3>=1.35.0",
  # Observability
  "prometheus-client>=0.20.0",
  "python-json-logger>=2.0.7",
  # UI
  "streamlit>=1.38.0",
  "requests>=2.32.3",
  "chardet<6",
]

[project.optional-dependencies]
paddle = [
  "paddleocr>=2.8.1",
]
ml = [
  "transformers>=4.44.2",
  "torch>=2.4.1",
]
dev = [
  "pytest>=8.3.2",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.4",
  "mypy>=1.11.0",
  "httpx>=0.27.2",
]

[tool.setuptools.packages.find]
include = ["app*", "review_ui*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
